chore: rename/refactor aws/session into aws/auth
