+++ pkg/provider/aws/auth/fake/assumeroler.go
